bugfixe

+++ spirl/rl/components/sampler.py
@@ -213,7 +213,7 @@ class HierarchicalSampler(Sampler):
 class ImageSampler(Sampler):
     """Appends image rendering to raw observation."""
     def _postprocess_obs(self, obs):
-        img = self._env.render().transpose(2, 0, 1).flatten()
+        img = self._env.render().transpose(2, 0, 1).flatten().astype(np.uint8)
         return img
 
 class MultiImageSampler(Sampler):
@@ -227,7 +227,7 @@ class MultiImageSampler(Sampler):
         if not self._past_frames:   # initialize past frames with N copies of current frame
             [self._past_frames.append(img) for _ in range(self._hp.n_frames - 1)]
         self._past_frames.append(img)
-        stacked_img = np.concatenate(list(self._past_frames), axis=0)
+        stacked_img = np.concatenate(list(self._past_frames), axis=0).astype(np.uint8)
         return stacked_img
 
 class ImageAugmentedSampler(Sampler):
